backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=2.0.0
python-dotenv==1.0.1
requests>=2.31.0
importlib-metadata
fasttext>=0.9.2
